requirements.txt
google-generativeai==0.5.2
pyttsx3==2.90
pillow==10.3.0
SpeechRecognition==3.10.4
PyAudio==0.2.14